final refactor
