+++ src/main/java/org/icgc/argo/workflow_management/graphql/model/GqlWorkflowEngineParams.java
@@ -16,23 +16,13 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc.argo.workflow_management.controller.model.wes;
+package org.icgc.argo.workflow_management.graphql.model;
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import io.swagger.annotations.ApiModel;
-import javax.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import org.icgc.argo.workflow_management.wes.controller.model.WorkflowEngineParams;
 
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-@ApiModel(
-    description =
-        "A successful run will return the runId which can be used to get info about that run")
-public class RunsResponse {
-  @NotNull private String runId;
+@JsonNaming()
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class GqlWorkflowEngineParams extends WorkflowEngineParams {
 }
