+++ src/main/java/org/icgc/argo/workflow_management/properties/Swagger2Properties.java
@@ -47,7 +47,7 @@ public class Swagger2Properties {
   public Docket api(SwaggerProperties properties) {
     return new Docket(DocumentationType.SWAGGER_2)
         .select()
-        .apis(RequestHandlerSelectors.basePackage("org.icgc.argo.workflow_management.controller"))
+        .apis(RequestHandlerSelectors.basePackage("org.icgc.argo.workflow_management.wes.controller"))
         .paths(PathSelectors.any())
         .build()
         .host(properties.host)
