+++ src/main/java/org/icgc/argo/workflow_management/service/WorkflowExecutionService.java
@@ -20,14 +20,14 @@ package org.icgc.argo.workflow_management.service;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import org.icgc.argo.workflow_management.controller.model.wes.RunsResponse;
-import org.icgc.argo.workflow_management.service.model.WESRunParams;
+import org.icgc.argo.workflow_management.wes.controller.model.RunsResponse;
+import org.icgc.argo.workflow_management.service.model.RunParams;
 import org.springframework.security.access.prepost.PreAuthorize;
 import reactor.core.publisher.Mono;
 
 public interface WorkflowExecutionService {
   @HasQueryAndMutationAccess
-  Mono<RunsResponse> run(WESRunParams params);
+  Mono<RunsResponse> run(RunParams params);
 
   @HasQueryAndMutationAccess
   Mono<RunsResponse> cancel(String runId);
