+++ src/main/java/org/icgc/argo/workflow_management/wes/controller/model/WorkflowEngineParams.java
@@ -16,20 +16,29 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc.argo.workflow_management.service.model;
+package org.icgc.argo.workflow_management.wes.controller.model;
 
-import java.util.Map;
-import lombok.Builder;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import io.swagger.annotations.ApiModel;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.icgc.argo.workflow_management.controller.model.wes.WorkflowEngineParams;
+import lombok.NoArgsConstructor;
+import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 
 @Data
-@Builder
-@RequiredArgsConstructor
-public class WESRunParams {
-  @NonNull private final Map<String, Object> workflowParams;
-  @NonNull private final String workflowUrl;
-  private final WorkflowEngineParams workflowEngineParams;
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+@JsonIgnoreProperties(ignoreUnknown = true)
+@ApiModel(description = "Describes valid workflow engine parameters (part of RunsRequest)")
+public class WorkflowEngineParams {
+  private String defaultContainer;
+  private String revision;
+  private UUID resume;
+  private String launchDir;
+  private String projectDir;
+  private String workDir;
+  private String latest;
 }
