+++ src/test/java/org/icgc/argo/workflow_management/util/WorkflowTest.java
@@ -21,11 +21,11 @@ package org.icgc.argo.workflow_management.util;
 import java.io.IOException;
 import java.util.Map;
 import lombok.val;
-import org.icgc.argo.workflow_management.controller.model.wes.WorkflowEngineParams;
+import org.icgc.argo.workflow_management.wes.controller.model.WorkflowEngineParams;
 import org.icgc.argo.workflow_management.secret.SecretProvider;
 import org.icgc.argo.workflow_management.secret.impl.NoSecretProvider;
 import org.icgc.argo.workflow_management.service.NextflowService;
-import org.icgc.argo.workflow_management.service.model.WESRunParams;
+import org.icgc.argo.workflow_management.service.model.RunParams;
 import org.icgc.argo.workflow_management.service.properties.NextflowProperties;
 
 public class WorkflowTest {
@@ -33,7 +33,7 @@ public class WorkflowTest {
   public static void main(String[] args) throws IOException {
     val p = new WorkflowEngineParams();
     val params =
-        WESRunParams.builder()
+        RunParams.builder()
             .workflowUrl("nextflow-io/hello")
             .workflowParams(Map.of())
             .workflowEngineParams(p)
@@ -41,7 +41,7 @@ public class WorkflowTest {
     runTest(params);
   }
 
-  static void runTest(WESRunParams params) {
+  static void runTest(RunParams params) {
     NextflowProperties config = new NextflowProperties();
     SecretProvider secretProvider = new NoSecretProvider();
     val service = new NextflowService(config, secretProvider);
